- parse BigNumber for create - Modify validation for create
