+++ src/components/handshakes/betting/utils.js
@@ -1,15 +1,15 @@
 import { MasterWallet } from '@/services/Wallets/MasterWallet';
 import { BettingHandshake } from '@/services/neuron';
 import local from '@/services/localStore';
-import moment from 'moment';
 import { APP } from '@/constants';
-import { BET_TYPE, MESSAGE_SERVER, MESSAGE } from '@/components/handshakes/betting/Feed/BetHandshakeHandler';
+import { BET_TYPE, MESSAGE_SERVER } from '@/components/handshakes/betting/Feed/BetHandshakeHandler';
+
 import {BigNumber} from 'bignumber.js';
 import _ from 'lodash';
 export const parseBigNumber = (value)=>{
   return new BigNumber(value);
 }
-export const getMessageWithCode= (code)=> {
+export const getMessageWithCode = (code)=> {
     const keys = Object.keys(MESSAGE_SERVER).filter(k => k == code); // ["A", "B"]
     console.log('Keys:', keys);
     const value = keys.map(k => MESSAGE_SERVER[k]); // [0, 1]
@@ -26,48 +26,6 @@ export const getAddress = () => {
   const chainId = getChainIdDefaultWallet();
   const bettinghandshake = new BettingHandshake(chainId);
   return bettinghandshake.address;
-};
-
-
-export const isSameAddress = (address) => {
-  const currentAddress = getAddress();
-  if (address !== currentAddress) {
-    return false;
-  }
-  return true;
-};
-
-export const isRightNetwork = () => {
-  const wallet = MasterWallet.getWalletDefault('ETH');
-  MasterWallet.log(MasterWallet.getWalletDefault('ETH'));
-
-  if (process.env.isStaging) {
-    return true;
-  }
-  if (process.env.isProduction) {
-    if (wallet.network === MasterWallet.ListCoin[wallet.className].Network.Mainnet) {
-      return true;
-    }
-    return false;
-  }
-  return true;
-};
-
-export const isExpiredDate = (reportTime) => {
-  // const newClosingDate = moment.unix(closingDate).add(90, 'minutes');
-  console.log('Report Time:', reportTime);
-  const newClosingDate = moment.unix(reportTime);
-  const dayUnit = newClosingDate.utc();
-  const today = moment();
-  const todayUnit = today.utc();
-  // console.log('Closing Unix:', closingDateUnit.format());
-  console.log('New Date Unix:', dayUnit.format());
-  console.log('Today Unix:', todayUnit.format());
-  if (!todayUnit.isSameOrBefore(dayUnit, 'miliseconds') && today) {
-    console.log('Expired Date');
-    return true;
-  }
-  return false;
 };
 
 export const getId = (idOffchain) => {
@@ -100,7 +58,7 @@ export const getEstimateGas = async () => {
 export const foundShakeList = (item, offchain) => {
   // const shakerList = [];
   const profile = local.get(APP.AUTH_PROFILE);
-  const { shakers, outcome_id, from_address } = item;
+  const { shakers } = item;
 
   const idOffchain = getId(offchain);
   console.log('Id Offchain:', idOffchain);
@@ -137,45 +95,3 @@ export const isInitBet = (dict) => {
   return false;
 };
 
-export const validateBet = async (amount = 0, odds = 0, closingDate, matchName = '', matchOutcome = '', freeBet=false) => {
-  const balance = await getBalance();
-  const estimateGas = await getEstimateGas();
-  const estimatedGasBN = parseBigNumber(estimateGas.toString()||0);
-  const total = amount.plus(estimatedGasBN).toNumber()||0;
-  let result = { status: true, message: '' };
-  if (!isRightNetwork()) {
-    result.message = MESSAGE.RIGHT_NETWORK;
-    result.status = false;
-    return result;
-  }
-
-  if (matchName.length === 0 || matchOutcome.length === 0) {
-    result.message = MESSAGE.CHOOSE_MATCH;
-    result.status = false;
-    return result;
-  }
-
-  if (isExpiredDate(closingDate)) {
-    result.message = MESSAGE.MATCH_OVER;
-    result.status = false;
-    return result;
-  }
-  if (amount <= 0) {
-    result.message = MESSAGE.AMOUNT_VALID;
-    result.status = false;
-    return result;
-  }
-
-  if (total > balance && !freeBet) {
-    result.message = MESSAGE.NOT_ENOUGH_BALANCE;
-    result.status = false;
-    return result;
-  }
-
-  if (odds <= 1 || odds >= 12) {
-    result.message = MESSAGE.ODD_LARGE_THAN;
-    result.status = false;
-    return result;
-  }
-  return result;
-};
